requirements-saas.txt
# SaaS-specific dependencies
Flask==3.0.0
Flask-Login==0.6.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
cryptography==41.0.7
Werkzeug==3.0.1
